[RENAME]: Enum 유효성 관련 클래스 패키지 이동
- common.exception -> common.validation
