+++ src/main/java/com/carrot/carrotmarketclonecoding/common/validation/EnumValidator.java
@@ -1,4 +1,4 @@
-package com.carrot.carrotmarketclonecoding.common.exception;
+package com.carrot.carrotmarketclonecoding.common.validation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
